refactor:에러코드 수정 및 안 쓰는 파라미터 제거 -#7
