+++ src/main/java/com/mju_lion/letter/service/AuthService.java
@@ -1,8 +1,8 @@
 package com.mju_lion.letter.service;
 
 import com.mju_lion.letter.authentication.PasswordHashEncryption;
-import com.mju_lion.letter.dto.LoginDto;
-import com.mju_lion.letter.dto.SignupDto;
+import com.mju_lion.letter.dto.request.auth.LoginDto;
+import com.mju_lion.letter.dto.request.auth.SignupDto;
 import com.mju_lion.letter.entity.User;
 import com.mju_lion.letter.error.ErrorCode;
 import com.mju_lion.letter.error.exception.ConflictException;
@@ -25,7 +25,7 @@ public class AuthService {
 
         User user=userRepository.findByUserId(signupDto.getUserId());
         if(null!=user){
-            throw new ConflictException(ErrorCode.USERID__ALREADY_EXISTS);
+            throw new ConflictException(ErrorCode.USERID_ALREADY_EXISTS);
         }
 
         //유저 만들어주기
